chore(processor): fix tests
