--- processor/test/services/converters/lineitem.converter.spec.ts
@@ -1,275 +1,0 @@
-import { describe, test, expect, afterEach, jest, beforeEach } from '@jest/globals';
-import {
-  mapCoCoCartItemsToAdyenLineItems,
-  mapCoCoLineItemToAdyenLineItem,
-  mapCoCoCustomLineItemToAdyenLineItem,
-  mapCoCoShippingInfoToAdyenLineItem,
-  mapCoCoDiscountOnTotalPriceToAdyenLineItem,
-} from '../../../src/services/converters/lineitem.converter';
-import { LineItem } from '@adyen/api-library/lib/src/typings/checkout/lineItem';
-import { Cart, LineItem as CoCoLineItem, CustomLineItem, ShippingInfo } from '@commercetools/connect-payments-sdk';
-import CoCoCartSimpleJSON from '../../data/coco-cart-simple-shipping.json';
-import CoCoCartMultipleJSON from '../../data/coco-cart-multiple-shipping.json';
-import CoCoCartCLPJSON from '../../data/coco-cart-clp.json';
-import CartDiscounts from '../../data/cart-discounts.json';
-
-describe('helper.converter', () => {
-  beforeEach(() => {
-    jest.setTimeout(10000);
-    jest.resetAllMocks();
-  });
-
-  afterEach(() => {
-    jest.restoreAllMocks();
-  });
-
-  test('should map the CoCo line items to Adyen line items', () => {
-    const input = CoCoCartSimpleJSON.lineItems as CoCoLineItem[];
-
-    const actual = mapCoCoLineItemToAdyenLineItem(input[0]);
-    const expected: LineItem[] = [
-      {
-        id: 'WCSI-09',
-        description: 'Walnut Counter Stool',
-        quantity: 1,
-        amountExcludingTax: 7562,
-        amountIncludingTax: 8999,
-        taxAmount: 1437,
-        taxPercentage: 1900,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map the CoCo custom line items to Adyen line items', () => {
-    const input: CustomLineItem[] = CoCoCartSimpleJSON.customLineItems as CustomLineItem[];
-
-    const actual = mapCoCoCustomLineItemToAdyenLineItem(input[0]);
-    const expected: LineItem[] = [
-      {
-        id: 'fbb6ea54-e96b-4ebe-b4c9-ccad345bd5d6',
-        description: 'Walnut Counter Stool',
-        quantity: 1,
-        amountExcludingTax: 7562,
-        amountIncludingTax: 8999,
-        taxAmount: 1437,
-        taxPercentage: 1900,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map CoCo shipping info to Adyen line item', () => {
-    const input: ShippingInfo = CoCoCartSimpleJSON.shippingInfo as ShippingInfo;
-
-    const actual = mapCoCoShippingInfoToAdyenLineItem([
-      { shippingInfo: input, shippingAddress: CoCoCartSimpleJSON.shippingAddress },
-    ]);
-    const expected: LineItem[] = [
-      {
-        description: 'Shipping - Standard Delivery',
-        quantity: 1,
-        amountExcludingTax: 1344,
-        amountIncludingTax: 1599,
-        taxAmount: 255,
-        taxPercentage: 1900,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map CoCo shipping info to Adyen line item when multiple shipping', () => {
-    const actual = mapCoCoShippingInfoToAdyenLineItem([
-      {
-        shippingInfo: CoCoCartMultipleJSON.shipping[0].shippingInfo as ShippingInfo,
-        shippingAddress: CoCoCartMultipleJSON.shipping[0].shippingAddress,
-      },
-      {
-        shippingInfo: CoCoCartMultipleJSON.shipping[1].shippingInfo as ShippingInfo,
-        shippingAddress: CoCoCartMultipleJSON.shipping[1].shippingAddress,
-      },
-    ]);
-    const expected: LineItem[] = [
-      {
-        amountExcludingTax: 9091,
-        amountIncludingTax: 10000,
-        description: 'Shipping - ddelizia-delivery',
-        quantity: 1,
-        taxAmount: 909,
-        taxPercentage: 1000,
-      },
-      {
-        amountExcludingTax: 870,
-        amountIncludingTax: 1000,
-        description: 'Shipping - Express Delivery',
-        quantity: 1,
-        taxAmount: 130,
-        taxPercentage: 1500,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map CoCo discount info to Adyen line item', () => {
-    const input = CoCoCartSimpleJSON.discountOnTotalPrice as any;
-
-    const actual = mapCoCoDiscountOnTotalPriceToAdyenLineItem({ discountOnTotalPrice: input });
-    const expected: LineItem = {
-      description: 'Discount',
-      quantity: 1,
-      amountExcludingTax: -891,
-      amountIncludingTax: -1060,
-      taxAmount: -169,
-    };
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map a CoCo cart to Adyen line items', () => {
-    const input = CoCoCartSimpleJSON as Cart;
-
-    const actual = mapCoCoCartItemsToAdyenLineItems(input);
-    const expected: LineItem[] = [
-      {
-        id: 'WCSI-09',
-        description: 'Walnut Counter Stool',
-        quantity: 1,
-        amountExcludingTax: 7562,
-        amountIncludingTax: 8999,
-        taxAmount: 1437,
-        taxPercentage: 1900,
-      },
-      {
-        id: 'fbb6ea54-e96b-4ebe-b4c9-ccad345bd5d6',
-        description: 'Walnut Counter Stool',
-        quantity: 1,
-        amountExcludingTax: 7562,
-        amountIncludingTax: 8999,
-        taxAmount: 1437,
-        taxPercentage: 1900,
-      },
-      {
-        description: 'Shipping - Standard Delivery',
-        quantity: 1,
-        amountExcludingTax: 1344,
-        amountIncludingTax: 1599,
-        taxAmount: 255,
-        taxPercentage: 1900,
-      },
-      {
-        description: 'Discount',
-        quantity: 1,
-        amountExcludingTax: -891,
-        amountIncludingTax: -1060,
-        taxAmount: -169,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map a CoCo cart to Adyen line items when shipping mode is Multiple', () => {
-    const input = CoCoCartMultipleJSON as Cart;
-
-    const actual = mapCoCoCartItemsToAdyenLineItems(input);
-    const expected: LineItem[] = [
-      {
-        amountExcludingTax: 817,
-        amountIncludingTax: 899,
-        description: 'Willow Teapot',
-        id: 'WTP-09',
-        quantity: 1,
-        taxAmount: 82,
-        taxPercentage: 0,
-      },
-      {
-        amountExcludingTax: 9091,
-        amountIncludingTax: 10000,
-        description: 'Shipping - ddelizia-delivery',
-        quantity: 1,
-        taxAmount: 909,
-        taxPercentage: 1000,
-      },
-      {
-        amountExcludingTax: 870,
-        amountIncludingTax: 1000,
-        description: 'Shipping - Express Delivery',
-        quantity: 1,
-        taxAmount: 130,
-        taxPercentage: 1500,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map the CoCo line items to Adyen line items taking into account Adyen deviations', () => {
-    // CLP currencyCode according to ISO_4217 has 0 fractionDigits but Adyen expects 2 fractionDigits
-    const input = CoCoCartCLPJSON.lineItems as CoCoLineItem[];
-
-    const actual = mapCoCoLineItemToAdyenLineItem(input[0]);
-    const expected: LineItem[] = [
-      {
-        id: 'TST-02',
-        description: 'Teak Serving Platter',
-        quantity: 1,
-        amountExcludingTax: 13400,
-        amountIncludingTax: 15000,
-        taxAmount: 1600,
-        taxPercentage: 1200,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-  });
-
-  test('should map a CoCo cart which contains product-discounts and cart-discounts-which-target-products and cart-discounts on totalPrice to Adyen line items', () => {
-    const input = CartDiscounts as Cart;
-
-    const actual = mapCoCoCartItemsToAdyenLineItems(input);
-    const expected: LineItem[] = [
-      {
-        id: 'MTSS-01',
-        description: 'Modern Three Seater Sofa',
-        quantity: 1,
-        amountExcludingTax: 142500,
-        amountIncludingTax: 171000,
-        taxAmount: 28500,
-        taxPercentage: 2000,
-      },
-      {
-        id: 'MTSS-01-discount',
-        description: 'Modern Three Seater Sofa discount',
-        quantity: 1,
-        amountIncludingTax: -29000,
-      },
-      {
-        description: 'Discount',
-        quantity: 1,
-        amountExcludingTax: -34200,
-        amountIncludingTax: -41040,
-        taxAmount: -6840,
-      },
-    ];
-
-    expect(actual).toEqual(expected);
-
-    //  41040 is the total discountOnTotalPrice (cart-discounts)
-    //  29000 is the total productDiscounts (including direct product-discounts + cart-discounts which target products)
-    //      +
-    //  70040 is the total centAmount discount from everything
-    //      +
-    // 129960 cart.totalPrice (user has to pay)
-    // 200000 equals the single line item original price without any discounts
-
-    const totalDiscounts = Math.abs(actual[1].amountIncludingTax ?? 0) + Math.abs(actual[2].amountIncludingTax ?? 0);
-    const totalValue = input.totalPrice.centAmount + totalDiscounts;
-
-    expect(input.lineItems).toHaveLength(1);
-    expect(totalValue).toEqual(input.lineItems[0].price.value.centAmount);
-  });
-});
